0110 자원 시장 구성 변경 트레이더 작동 확인

+++ Assets/Scripts/Data/Resource/ResourceData.cs
@@ -6,7 +6,8 @@ public class ResourceData : ScriptableObject
 {
     public string id;
     public string displayName;
-    public float baseValue = 10f;
+
+    public long baseValue = 10;
 
     public ResourceType type;
     public Sprite icon;
